+++ crates/dispatch2/src/object.rs
@@ -1,16 +1,10 @@
-//! Dispatch object definition.
+use core::ptr::NonNull;
 
 use alloc::boxed::Box;
 
-use super::{ffi::*, queue::Queue, utils::function_wrapper, QualityOfServiceClass};
+use crate::DispatchRetained;
 
-/// Error returned by [DispatchObject::set_target_queue].
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
-#[non_exhaustive]
-pub enum TargetQueueError {
-    /// The [DispatchObject] is already active.
-    ObjectAlreadyActive,
-}
+use super::{ffi::*, queue::DispatchQueue, utils::function_wrapper, QualityOfServiceClass};
 
 /// Error returned by [DispatchObject::set_qos_class_floor].
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
@@ -20,48 +14,28 @@ pub enum QualityOfServiceClassFloorError {
     InvalidRelativePriority,
 }
 
-/// Represent a dispatch object.
-#[repr(C)]
-#[derive(Debug)]
-pub struct DispatchObject<T> {
-    object: *mut T,
-    is_activated: bool,
-}
-
-impl<T> DispatchObject<T> {
-    /// Create a new owned instance
+/// Types that represent dispatch objects.
+///
+/// # Safety
+///
+/// The object must represent a dispatch object, and be usable in
+/// `dispatch_retain` / `dispatch_release`.
+pub unsafe trait DispatchObject {
+    /// Increment the reference count of the object.
     ///
-    /// # Safety
-    ///
-    /// - ``object`` is expected to be a dispatch object that is owned.
-    pub unsafe fn new_owned(object: *mut T) -> Self {
-        Self {
-            object,
-            is_activated: false,
-        }
-    }
-
-    /// Create a new shared instance
-    ///
-    /// # Safety
-    ///
-    /// - ``object`` is expected to be a dispatch object that is shared.
-    pub unsafe fn new_shared(object: *mut T) -> Self {
-        let result = Self {
-            object,
-            is_activated: false,
-        };
-
-        // Safety: We own a reference to the object.
-        unsafe {
-            dispatch_retain(result.object.cast());
-        }
-
-        result
+    /// This extends the duration in which the object is alive by detaching it
+    /// from the lifetime information carried by the reference.
+    fn retain(&self) -> DispatchRetained<Self> {
+        let ptr: NonNull<Self> = NonNull::from(self);
+        // SAFETY:
+        // - The pointer is valid since it came from `&self`.
+        // - The lifetime of the pointer itself is extended, but any lifetime
+        //   that the object may carry is still kept within the type itself.
+        unsafe { DispatchRetained::retain(ptr) }
     }
 
     /// Set the finalizer function for the object.
-    pub fn set_finalizer<F>(&mut self, destructor: F)
+    fn set_finalizer<F>(&mut self, destructor: F)
     where
         F: Send + FnOnce(),
     {
@@ -71,27 +45,24 @@ impl<T> DispatchObject<T> {
         //         Once the finalizer is executed, the context will be dangling.
         //         This isn't an issue as the context shall not be accessed after the dispatch object is destroyed.
         unsafe {
-            dispatch_set_context(self.object.cast(), destructor_boxed);
-            dispatch_set_finalizer_f(self.object.cast(), function_wrapper::<F>)
+            dispatch_set_context(self.as_raw(), destructor_boxed);
+            dispatch_set_finalizer_f(self.as_raw(), function_wrapper::<F>)
         }
     }
 
-    /// Set the target [Queue] of this object.
+    /// Set the target [`DispatchQueue`] of this object.
+    ///
+    /// # Aborts
+    ///
+    /// Aborts if the object has been activated.
     ///
     /// # Safety
     ///
-    /// - DispatchObject should be a queue or queue source.
-    pub unsafe fn set_target_queue(&self, queue: &Queue) -> Result<(), TargetQueueError> {
-        if self.is_activated {
-            return Err(TargetQueueError::ObjectAlreadyActive);
-        }
-
-        // SAFETY: object and queue cannot be null.
-        unsafe {
-            dispatch_set_target_queue(self.as_raw().cast(), queue.as_raw());
-        }
-
-        Ok(())
+    /// - There must not be a cycle in the hierarchy of queues.
+    #[doc(alias = "dispatch_set_target_queue")]
+    unsafe fn set_target_queue(&self, queue: &DispatchQueue) {
+        // SAFETY: `object` and `queue` cannot be null, rest is upheld by caller.
+        unsafe { dispatch_set_target_queue(self.as_raw(), queue.as_raw()) };
     }
 
     /// Set the QOS class floor on a dispatch queue, source or workloop.
@@ -99,7 +70,7 @@ impl<T> DispatchObject<T> {
     /// # Safety
     ///
     /// - DispatchObject should be a queue or queue source.
-    pub unsafe fn set_qos_class_floor(
+    unsafe fn set_qos_class_floor(
         &self,
         qos_class: QualityOfServiceClass,
         relative_priority: i32,
@@ -111,7 +82,7 @@ impl<T> DispatchObject<T> {
         // SAFETY: Safe as relative_priority can only be valid.
         unsafe {
             dispatch_set_qos_class_floor(
-                self.as_raw().cast(),
+                self.as_raw(),
                 dispatch_qos_class_t::from(qos_class),
                 relative_priority,
             );
@@ -121,53 +92,26 @@ impl<T> DispatchObject<T> {
     }
 
     /// Activate the object.
-    pub fn activate(&mut self) {
+    fn activate(&mut self) {
         // Safety: object cannot be null.
-        unsafe {
-            dispatch_activate(self.as_raw().cast());
-        }
-
-        self.is_activated = true;
+        unsafe { dispatch_activate(self.as_raw()) };
     }
 
     /// Suspend the invocation of functions on the object.
-    pub fn suspend(&self) {
+    fn suspend(&self) {
         // Safety: object cannot be null.
-        unsafe {
-            dispatch_suspend(self.as_raw().cast());
-        }
+        unsafe { dispatch_suspend(self.as_raw()) };
     }
 
     /// Resume the invocation of functions on the object.
-    pub fn resume(&self) {
+    fn resume(&self) {
         // Safety: object cannot be null.
-        unsafe {
-            dispatch_resume(self.as_raw().cast());
-        }
+        unsafe { dispatch_resume(self.as_raw()) };
     }
 
-    /// Get the raw object value.
-    ///
-    /// # Safety
-    ///
-    /// - Object shouldn't be released manually.
-    pub const unsafe fn as_raw(&self) -> *mut T {
-        self.object
-    }
-}
-
-impl<T> Clone for DispatchObject<T> {
-    fn clone(&self) -> Self {
-        // Safety: We own a reference to the object.
-        unsafe { Self::new_shared(self.object) }
-    }
-}
-
-impl<T> Drop for DispatchObject<T> {
-    fn drop(&mut self) {
-        // Safety: We own a reference to the object.
-        unsafe {
-            dispatch_release(self.object.cast());
-        }
+    #[doc(hidden)]
+    fn as_raw(&self) -> dispatch_object_t {
+        let ptr: *const Self = self;
+        ptr as dispatch_object_t
     }
 }