+++ crates/dispatch2/src/queue.rs
@@ -1,17 +1,15 @@
-//! Dispatch queue definition.
-
 use alloc::boxed::Box;
 use alloc::ffi::CString;
-use core::borrow::{Borrow, BorrowMut};
-use core::ops::{Deref, DerefMut};
 use core::ptr::NonNull;
 use core::time::Duration;
 
-use super::object::{DispatchObject, QualityOfServiceClassFloorError, TargetQueueError};
+use crate::DispatchRetained;
+
+use super::object::{DispatchObject, QualityOfServiceClassFloorError};
 use super::utils::function_wrapper;
 use super::{ffi::*, QualityOfServiceClass};
 
-/// Error returned by [Queue::after].
+/// Error returned by [`DispatchQueue::after`].
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
 #[non_exhaustive]
 pub enum QueueAfterError {
@@ -67,7 +65,7 @@ impl From<QueuePriority> for dispatch_queue_priority_t {
     }
 }
 
-/// Global queue identifier definition for [Queue::new] and [Queue::new_with_target].
+/// Global queue identifier definition for [`DispatchQueue::new`] and [`DispatchQueue::new_with_target`].
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub enum GlobalQueueIdentifier {
     /// Standard priority based queue.
@@ -90,45 +88,14 @@ impl GlobalQueueIdentifier {
     }
 }
 
-/// Auto release frequency for [WorkloopQueue::set_autorelease_frequency].
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
-#[non_exhaustive]
-pub enum DispatchAutoReleaseFrequency {
-    /// Inherit autorelease frequency from the target [Queue].
-    Inherit,
-    /// Configure an autorelease pool before the execution of a function and releases the objects in that pool after the function finishes executing.
-    WorkItem,
-    /// Never setup an autorelease pool.
-    Never,
-}
+dispatch_object!(
+    /// Dispatch queue.
+    pub struct DispatchQueue;
+);
 
-impl From<DispatchAutoReleaseFrequency> for dispatch_autorelease_frequency_t {
-    fn from(value: DispatchAutoReleaseFrequency) -> Self {
-        match value {
-            DispatchAutoReleaseFrequency::Inherit => {
-                dispatch_autorelease_frequency_t::DISPATCH_AUTORELEASE_FREQUENCY_INHERIT
-            }
-            DispatchAutoReleaseFrequency::WorkItem => {
-                dispatch_autorelease_frequency_t::DISPATCH_AUTORELEASE_FREQUENCY_WORK_ITEM
-            }
-            DispatchAutoReleaseFrequency::Never => {
-                dispatch_autorelease_frequency_t::DISPATCH_AUTORELEASE_FREQUENCY_NEVER
-            }
-            _ => panic!("Unknown DispatchAutoReleaseFrequency value: {:?}", value),
-        }
-    }
-}
-
-/// Dispatch queue.
-#[derive(Debug, Clone)]
-pub struct Queue {
-    dispatch_object: DispatchObject<dispatch_queue_s>,
-    is_workloop: bool,
-}
-
-impl Queue {
-    /// Create a new [Queue].
-    pub fn new(label: &str, queue_attribute: QueueAttribute) -> Self {
+impl DispatchQueue {
+    /// Create a new [`DispatchQueue`].
+    pub fn new(label: &str, queue_attribute: QueueAttribute) -> DispatchRetained<Self> {
         let label = CString::new(label).expect("Invalid label!");
 
         // Safety: label and queue_attribute can only be valid.
@@ -136,94 +103,72 @@ impl Queue {
             dispatch_queue_create(label.as_ptr(), dispatch_queue_attr_t::from(queue_attribute))
         };
 
-        assert!(!object.is_null(), "dispatch_queue_create shouldn't fail!");
-
-        // Safety: object cannot be null.
-        let dispatch_object = unsafe { DispatchObject::new_owned(object.cast()) };
-
-        Queue {
-            dispatch_object,
-            is_workloop: false,
-        }
+        let object = NonNull::new(object).expect("dispatch_queue_create returned NULL");
+        // SAFETY: Object came from a "create" method.
+        unsafe { DispatchRetained::from_raw(object.cast()) }
     }
 
-    /// Create a new [Queue] with a given target [Queue].
-    pub fn new_with_target(label: &str, queue_attribute: QueueAttribute, target: &Queue) -> Self {
+    /// Create a new [`DispatchQueue`] with a given target [`DispatchQueue`].
+    pub fn new_with_target(
+        label: &str,
+        queue_attribute: QueueAttribute,
+        target: &DispatchQueue,
+    ) -> DispatchRetained<Self> {
         let label = CString::new(label).expect("Invalid label!");
 
         // Safety: label, queue_attribute and target can only be valid.
+        // NOTE: dispatch_queue_create_with_target is in charge of retaining the target DispatchQueue.
         let object = unsafe {
             dispatch_queue_create_with_target(
                 label.as_ptr(),
                 dispatch_queue_attr_t::from(queue_attribute),
-                target.dispatch_object.as_raw(),
+                target.as_raw(),
             )
         };
 
-        assert!(!object.is_null(), "dispatch_queue_create shouldn't fail!");
-
-        // Safety: object cannot be null.
-        let dispatch_object = unsafe { DispatchObject::new_owned(object.cast()) };
-
-        // NOTE: dispatch_queue_create_with_target is in charge of retaining the target Queue.
-
-        Queue {
-            dispatch_object,
-            is_workloop: false,
-        }
+        let object = NonNull::new(object).expect("dispatch_queue_create_with_target returned NULL");
+        // SAFETY: Object came from a "create" method.
+        unsafe { DispatchRetained::from_raw(object.cast()) }
     }
 
-    /// Return a system-defined global concurrent [Queue] with the priority derived from [GlobalQueueIdentifier].
-    pub fn global_queue(identifier: GlobalQueueIdentifier) -> Self {
+    /// Return a system-defined global concurrent [`DispatchQueue`] with the priority derived from [GlobalQueueIdentifier].
+    pub fn global_queue(identifier: GlobalQueueIdentifier) -> DispatchRetained<Self> {
         let raw_identifier = identifier.to_identifier();
 
         // Safety: raw_identifier cannot be invalid, flags is reserved.
         let object = unsafe { dispatch_get_global_queue(raw_identifier, 0) };
 
-        assert!(
-            !object.is_null(),
-            "dispatch_get_global_queue shouldn't fail!"
-        );
-
-        // Safety: object cannot be null.
-        let dispatch_object = unsafe { DispatchObject::new_shared(object.cast()) };
-
-        Queue {
-            dispatch_object,
-            is_workloop: false,
-        }
+        let object = NonNull::new(object).expect("dispatch_get_global_queue returned NULL");
+        // SAFETY: Object came from a "create" method.
+        unsafe { DispatchRetained::from_raw(object.cast()) }
     }
 
     /// Return the main queue.
-    pub fn main() -> Self {
+    pub fn main() -> DispatchRetained<Self> {
         // Safety: raw_identifier cannot be invalid, flags is reserved.
         let object = dispatch_get_main_queue();
 
-        assert!(!object.is_null(), "dispatch_get_main_queue shouldn't fail!");
-
-        // Safety: object cannot be null.
-        let dispatch_object = unsafe { DispatchObject::new_shared(object.cast()) };
-
-        Queue {
-            dispatch_object,
-            is_workloop: false,
-        }
+        let object = NonNull::new(object).expect("dispatch_get_main_queue returned NULL");
+        // SAFETY: Object came from a "create" method.
+        unsafe { DispatchRetained::from_raw(object.cast()) }
     }
 
-    /// Submit a function for synchronous execution on the [Queue].
+    /// Submit a function for synchronous execution on the [`DispatchQueue`].
     pub fn exec_sync<F>(&self, work: F)
     where
         F: Send + FnOnce(),
     {
-        assert!(!self.is_workloop, "exec_sync is invalid for WorkloopQueue");
-
         let work_boxed = Box::into_raw(Box::new(work)).cast();
 
+        // NOTE: `dispatch_sync*` functions are discouraged on workloops for
+        // performance reasons, but they should still work, so we won't forbid
+        // it here.
+        //
         // Safety: object cannot be null and work is wrapped to avoid ABI incompatibility.
         unsafe { dispatch_sync_f(self.as_raw(), work_boxed, function_wrapper::<F>) }
     }
 
-    /// Submit a function for asynchronous execution on the [Queue].
+    /// Submit a function for asynchronous execution on the [`DispatchQueue`].
     pub fn exec_async<F>(&self, work: F)
     where
         // We need `'static` to make sure any referenced values are borrowed for
@@ -236,7 +181,7 @@ impl Queue {
         unsafe { dispatch_async_f(self.as_raw(), work_boxed, function_wrapper::<F>) }
     }
 
-    /// Enqueue a function for execution at the specified time on the [Queue].
+    /// Enqueue a function for execution at the specified time on the [`DispatchQueue`].
     pub fn after<F>(&self, wait_time: Duration, work: F) -> Result<(), QueueAfterError>
     where
         F: Send + FnOnce(),
@@ -253,7 +198,7 @@ impl Queue {
         Ok(())
     }
 
-    /// Enqueue a barrier function for asynchronous execution on the [Queue] and return immediately.
+    /// Enqueue a barrier function for asynchronous execution on the [`DispatchQueue`] and return immediately.
     pub fn barrier_async<F>(&self, work: F)
     where
         // We need `'static` to make sure any referenced values are borrowed for
@@ -266,7 +211,7 @@ impl Queue {
         unsafe { dispatch_barrier_async_f(self.as_raw(), work_boxed, function_wrapper::<F>) }
     }
 
-    /// Enqueue a barrier function for synchronous execution on the [Queue] and wait until that function completes.
+    /// Enqueue a barrier function for synchronous execution on the [`DispatchQueue`] and wait until that function completes.
     pub fn barrier_sync<F>(&self, work: F)
     where
         F: Send + FnOnce(),
@@ -292,7 +237,7 @@ impl Queue {
         }
     }
 
-    /// Sets a function at the given key that will be executed at [Queue] destruction.
+    /// Sets a function at the given key that will be executed at [`DispatchQueue`] destruction.
     pub fn set_specific<F>(&mut self, key: NonNull<()>, destructor: F)
     where
         F: Send + FnOnce(),
@@ -314,46 +259,14 @@ impl Queue {
         }
     }
 
-    /// Set the finalizer function for the [Queue].
-    pub fn set_finalizer<F>(&mut self, destructor: F)
-    where
-        F: Send + FnOnce(),
-    {
-        self.dispatch_object.set_finalizer(destructor);
-    }
-
-    /// Set the target [Queue] of this [Queue].
-    pub fn set_target_queue(&self, queue: &Queue) -> Result<(), TargetQueueError> {
-        // Safety: We are in Queue instance.
-        unsafe { self.dispatch_object.set_target_queue(queue) }
-    }
-
-    /// Set the QOS class floor of the [Queue].
+    /// Set the QOS class floor of the [`DispatchQueue`].
     pub fn set_qos_class_floor(
         &self,
         qos_class: QualityOfServiceClass,
         relative_priority: i32,
     ) -> Result<(), QualityOfServiceClassFloorError> {
-        // Safety: We are in Queue instance.
-        unsafe {
-            self.dispatch_object
-                .set_qos_class_floor(qos_class, relative_priority)
-        }
-    }
-
-    /// Activate the [Queue].
-    pub fn activate(&mut self) {
-        self.dispatch_object.activate();
-    }
-
-    /// Suspend the invocation of functions on the [Queue].
-    pub fn suspend(&self) {
-        self.dispatch_object.suspend();
-    }
-
-    /// Resume the invocation of functions on the [Queue].
-    pub fn resume(&self) {
-        self.dispatch_object.resume();
+        // SAFETY: We are a queue.
+        unsafe { DispatchObject::set_qos_class_floor(self, qos_class, relative_priority) }
     }
 
     /// Get the raw [dispatch_queue_t] value.
@@ -362,106 +275,7 @@ impl Queue {
     ///
     /// - Object shouldn't be released manually.
     pub const unsafe fn as_raw(&self) -> dispatch_queue_t {
-        // SAFETY: Upheld by caller.
-        unsafe { self.dispatch_object.as_raw() }
-    }
-}
-
-/// Dispatch workloop queue.
-#[derive(Debug, Clone)]
-pub struct WorkloopQueue {
-    queue: Queue,
-}
-
-impl WorkloopQueue {
-    /// Create a new [WorkloopQueue].
-    pub fn new(label: &str, inactive: bool) -> Self {
-        let label = CString::new(label).expect("Invalid label!");
-
-        // Safety: label can only be valid.
-        let object = unsafe {
-            if inactive {
-                dispatch_workloop_create_inactive(label.as_ptr())
-            } else {
-                dispatch_workloop_create(label.as_ptr())
-            }
-        };
-
-        assert!(!object.is_null(), "dispatch_queue_create shouldn't fail!");
-
-        // Safety: object cannot be null.
-        let dispatch_object = unsafe { DispatchObject::new_owned(object.cast()) };
-
-        WorkloopQueue {
-            queue: Queue {
-                dispatch_object,
-                is_workloop: true,
-            },
-        }
-    }
-
-    /// Configure how the [WorkloopQueue] manage the autorelease pools for the functions it executes.
-    pub fn set_autorelease_frequency(&self, frequency: DispatchAutoReleaseFrequency) {
-        // Safety: object and frequency can only be valid.
-        unsafe {
-            dispatch_workloop_set_autorelease_frequency(
-                self.as_raw(),
-                dispatch_autorelease_frequency_t::from(frequency),
-            );
-        }
-    }
-
-    /// Get the raw [dispatch_workloop_t] value.
-    ///
-    /// # Safety
-    ///
-    /// - Object shouldn't be released manually.
-    pub const unsafe fn as_raw(&self) -> dispatch_workloop_t {
-        // SAFETY: Upheld by caller.
-        unsafe { self.queue.as_raw() as dispatch_workloop_t }
-    }
-}
-
-impl Deref for WorkloopQueue {
-    type Target = Queue;
-
-    #[inline]
-    fn deref(&self) -> &Self::Target {
-        &self.queue
-    }
-}
-
-impl DerefMut for WorkloopQueue {
-    #[inline]
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.queue
-    }
-}
-
-impl AsRef<Queue> for WorkloopQueue {
-    #[inline]
-    fn as_ref(&self) -> &Queue {
-        self
-    }
-}
-
-impl AsMut<Queue> for WorkloopQueue {
-    #[inline]
-    fn as_mut(&mut self) -> &mut Queue {
-        &mut *self
-    }
-}
-
-impl Borrow<Queue> for WorkloopQueue {
-    #[inline]
-    fn borrow(&self) -> &Queue {
-        self
-    }
-}
-
-impl BorrowMut<Queue> for WorkloopQueue {
-    #[inline]
-    fn borrow_mut(&mut self) -> &mut Queue {
-        &mut *self
+        let ptr: *const Self = self;
+        ptr as dispatch_queue_t
     }
 }